# docs/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/categories": {
            "get": {
                "operationId": "CategoriesController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Les cat\u00e9gories disponibles (Travail, Personnel, Courses, Sant\u00e9)",
                "tags": [
                    "categories"
                ]
            }
        },
        "/health": {
            "get": {
                "operationId": "HealthController_check",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "\u00c9tat de l\u2019API et de la base de donn\u00e9es",
                "tags": [
                    "health"
                ]
            }
        },
        "/auth/register": {
            "post": {
                "operationId": "AuthController_register",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "summary": "Cr\u00e9er un compte (un code OTP est envoy\u00e9 par e-mail)",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/verify-otp": {
            "post": {
                "operationId": "AuthController_verifyOtp",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOtpDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "V\u00e9rifier l\u2019adresse e-mail avec le code re\u00e7u",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/resend-otp": {
            "post": {
                "operationId": "AuthController_resendOtp",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResendOtpDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Renvoyer un code de v\u00e9rification",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "parameters": [
                    {
                        "name": "X-Token-Transport",
                        "in": "header",
                        "description": "\u00ab cookie \u00bb (clients web) : le refresh token est plac\u00e9 dans un cookie HttpOnly et absent du corps. \u00ab body \u00bb (d\u00e9faut, clients mobiles) : le refresh token est renvoy\u00e9 dans le corps JSON.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Se connecter : jeton d\u2019acc\u00e8s (15 min) + refresh token (7 j)",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "operationId": "AuthController_refresh",
                "parameters": [
                    {
                        "name": "X-Token-Transport",
                        "in": "header",
                        "description": "\u00ab cookie \u00bb (clients web) : le refresh token est plac\u00e9 dans un cookie HttpOnly et absent du corps. \u00ab body \u00bb (d\u00e9faut, clients mobiles) : le refresh token est renvoy\u00e9 dans le corps JSON.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Obtenir un nouveau jeton d\u2019acc\u00e8s (rotation du refresh token)",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "operationId": "AuthController_logout",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "summary": "Se d\u00e9connecter : r\u00e9voque la session et efface le cookie",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/me": {
            "get": {
                "operationId": "AuthController_me",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "L\u2019utilisateur connect\u00e9",
                "tags": [
                    "auth"
                ]
            }
        },
        "/tasks": {
            "get": {
                "operationId": "TasksController_findAll",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "default": 1,
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/Object"
                                }
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "maximum": 100,
                            "default": 10,
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/Object"
                                }
                            ]
                        }
                    },
                    {
                        "name": "done",
                        "required": false,
                        "in": "query",
                        "description": "Filtrer sur l\u2019\u00e9tat termin\u00e9",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "categoryId",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "priority",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "LOW",
                                "MEDIUM",
                                "HIGH"
                            ]
                        }
                    },
                    {
                        "name": "q",
                        "required": false,
                        "in": "query",
                        "description": "Recherche dans le titre (insensible \u00e0 la casse)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "default": "createdAt",
                            "type": "string",
                            "enum": [
                                "createdAt",
                                "updatedAt",
                                "dueDate",
                                "priority",
                                "title",
                                "id"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "required": false,
                        "in": "query",
                        "description": "Sens du tri (compatible avec le cours NestJS)",
                        "schema": {
                            "default": "desc",
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "clientId",
                        "required": false,
                        "in": "query",
                        "description": "Retrouver une t\u00e2che par son identifiant client",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Lister ses t\u00e2ches (pagin\u00e9, filtrable, triable)",
                "tags": [
                    "tasks"
                ]
            },
            "post": {
                "operationId": "TasksController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Cr\u00e9er une t\u00e2che (idempotent si clientId est fourni)",
                "tags": [
                    "tasks"
                ]
            }
        },
        "/tasks/{id}": {
            "get": {
                "operationId": "TasksController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Une t\u00e2che",
                "tags": [
                    "tasks"
                ]
            },
            "patch": {
                "operationId": "TasksController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Modifier une t\u00e2che (partiellement)",
                "tags": [
                    "tasks"
                ]
            },
            "delete": {
                "operationId": "TasksController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Supprimer une t\u00e2che (renvoie la t\u00e2che supprim\u00e9e)",
                "tags": [
                    "tasks"
                ]
            }
        }
    },
    "info": {
        "title": "TaskyAPI",
        "description": "API de r\u00e9f\u00e9rence du cours \u00ab Tasky Web : De Flutter au Web \u00bb",
        "version": "1.0.0",
        "contact": {}
    },
    "tags": [],
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "RegisterDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Awa Dossou"
                    },
                    "email": {
                        "type": "string",
                        "example": "awa@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "MotDePasse2026",
                        "minLength": 8
                    }
                },
                "required": [
                    "name",
                    "email",
                    "password"
                ]
            },
            "VerifyOtpDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "awa@example.com"
                    },
                    "code": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "required": [
                    "email",
                    "code"
                ]
            },
            "ResendOtpDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "awa@example.com"
                    }
                },
                "required": [
                    "email"
                ]
            },
            "LoginDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "demo@tasky.dev"
                    },
                    "password": {
                        "type": "string",
                        "example": "Tasky2026!"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "RefreshDto": {
                "type": "object",
                "properties": {
                    "refresh_token": {
                        "type": "string",
                        "description": "Mode \u00ab body \u00bb uniquement. En mode cookie, le jeton est lu dans le cookie tasky_refresh."
                    }
                }
            },
            "Object": {
                "type": "object",
                "properties": {}
            },
            "CreateTaskDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Acheter du pain",
                        "maxLength": 200
                    },
                    "description": {
                        "type": "object",
                        "example": "Pain complet, \u00e0 la boulangerie du coin",
                        "maxLength": 2000,
                        "nullable": true
                    },
                    "dueDate": {
                        "type": "object",
                        "example": "2026-09-30",
                        "description": "Date au format AAAA-MM-JJ",
                        "nullable": true
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                        ],
                        "default": "MEDIUM"
                    },
                    "categoryId": {
                        "type": "object",
                        "example": 3,
                        "nullable": true
                    },
                    "clientId": {
                        "type": "string",
                        "description": "Identifiant g\u00e9n\u00e9r\u00e9 par le client (UUID), pour le mode hors ligne"
                    },
                    "done": {
                        "type": "boolean",
                        "default": false
                    }
                },
                "required": [
                    "title"
                ]
            },
            "UpdateTaskDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Acheter du pain",
                        "maxLength": 200
                    },
                    "description": {
                        "type": "object",
                        "example": "Pain complet, \u00e0 la boulangerie du coin",
                        "maxLength": 2000,
                        "nullable": true
                    },
                    "dueDate": {
                        "type": "object",
                        "example": "2026-09-30",
                        "description": "Date au format AAAA-MM-JJ",
                        "nullable": true
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                        ],
                        "default": "MEDIUM"
                    },
                    "categoryId": {
                        "type": "object",
                        "example": 3,
                        "nullable": true
                    },
                    "clientId": {
                        "type": "string",
                        "description": "Identifiant g\u00e9n\u00e9r\u00e9 par le client (UUID), pour le mode hors ligne"
                    },
                    "done": {
                        "type": "boolean",
                        "default": false
                    }
                }
            }
        }
    }
}